refactor(onboarding): Create DegreePickerGallery component
Moved the implementation of DegreePicker from Onboarding_Pages/pg_1.tsx to
components/onboarding/DegreePickerGallery.tsx to reduce the complexity of
Onboarding_Pages/pg_1.tsx.
